DeskTop: When opening/refreshing a window, mark folders as open as needed.
DeskTop maintains a mapping of directory (volume/folder) icons to the
windows they are associated wit. In some cases (e.g. a window being
closed then re-opened, or a window refreshing) the association can get
lost, causing the folder to not be marked as open and other side
effects.

Recreate the association when populating a window with icons. This is
a prerequisite for fixing #594, but more work is needed.