DeskTop: Fix updating when file dragged into folder
When a file is dragged into a folder, if the folder is open then the
folder's window should be updated. But since the containing window is
rebuilt to represent the file being removed, the target folder's id is
no longer valid when the update is attempted.

Fix this by stashing the folder name before the update, and updating
the target id after the update. Fixes #594