File Pickers: Pull implementations/resources into lib/ #608
Dedupe the code, to prevent implementations from diverging.

DeskTop defines FD_EXTENDED as true, which currently flips on
all the changes. That's a big hammer, controlling all of:

* Hookable handler table
* Resources are in aux
* Two input fields supported
* Minor implementation differences

This can be made more granular as needed, and the minor
implementation differences can be removed over time.

No binary changes in this commit.