MGTK: Change MouseKeys exit to Escape key
Since SA is the mouse button in MouseKeys mode, requiring OA+SA+Space
to exit the mode was problematic. If SA was pressed first, it would be
treated as a button down. This could e.g. initiate a window drag,
which would be left active when the mode was exited.

So it's safest to have a separate sequence to exit the mode, and
Escape seems sufficient with no modifier required.