MGTK: Don't try to process mouse menu keys if not in menu mode.
When processing keyboard mouse movements, if arrows were not consumed
then the logic would fall unconditionally into menu key processing. If
Return was pressed, this would attempt to activate a menu item, and go
off the rails. In DeskTop, if an icon was selected this would cause
the cursor to become corrupted and exit explicit mousekeys mode.

Guard the logic by testing the specific mouse keys mode.