DeskTop/MGTK: Fix double-clicking flakiness. Fixes #628
(1) PeekEvent will return no_event if there's nothing in the queue.
641ae52 changed the DetectDoubleClick code to consume events after
peeking, but there's nothing in the queue. So that would wait for the
next event to appear, and throw it away, so a down event might get
missed. So a no_event *must not be* consumed.

(2) 641ae52 causes a real event no_event to be generated for MK mode
when the mouse isn't down (previously, it was always an up). This
no_event is really in the queue, and *must be* consumed or clicking
doesn't work in MK mode.

These are obviously incompatible. :)

The fix:

* DetectDoubleClick is changed to not consume a no_event.
* The MGTK change in 641ae52 is revised to not queue a no_event.