DeskTop: Support double-modifier shortcuts for open/close.
Support OA+SA+O for open-then-close-parent. Resolves #625.

Support OA+SA+W for close all. Resolves #626. Note that
a shortcut is not shown on the menu since MGTK doesn't support
multi-character shortcuts in menus, and it's not worth it to
add it.

Required tweaking MGTK's MouseKeys activation to not wait while
both modifiers are down, as that would stall events. Seems to
work just fine, though.