DeskTop: Handle operations with contained forked files. For #631
The initial fix for #631 handled operations directly on forked files.
This extends the fix to handle copy and delete traversals encountering
forked files.

Note that the behavior for delete is not ideal - since the containing
folder can't be deleted, it will appear as if locked. While
potentially annoying, this is not worth fixing - go use GS/OS instead.