DeskTop: Force windows to update after partial operation. Fixes #644
For copy/move/delete operations, if there is partial success (e.g.
some files were altered) then the source and (if needed) destination
windows need to be refreshed, but they were not if the operation
partially failed.

Switch the logic, and introduce a way to signal that an operation
(specifically, delete) was explicitly canceled before anything
happened.

Note that this does not use a "dirty flag", so in case of a failure
updates may occur that are not necessary, but this should be rare.