MGTK: Ensure when SetZP1 toggles to preserve that state is saved
The logic in SetZP1 when going from no-preserve (#$00) to preserve
(#$80) jumped to Dispatch::cleanup explicitly because the state
($F4-$FF) needs to be saved. But when that routine RTS'd it would...
run the same code again, trashing the saved state. The side effect was
that memory locations like $06-$0A would be overwritten by garbage.