DeskTop & Selector: Turn off MGTK ZP preservation to boost perf
Use MGTK's SetZP1 mechanism to turn off preservation of the zero page
for each MGTK call. This increases performance significantly
(repainting a window full of icons is about 25% faster). This is
applied to both DeskTop and Selector.

The base double-click and insertion point blink speeds are adjusted to
handle the increased performance; you'll probably want to re-check
your preferred settings in the Control Panel DA.

Zero page preservation is toggled back on when DAs are running. This
is to make them easier to program. Specifically, DAs like Eyes and
Calculator make use of the zero page for floating point operations.