Rename Startup DA file to Startup.Options. Fixes #649
BASIC.SYSTEM tries to invoke a BASIC file named STARTUP when launched.
When the Run.Basic.Here DA invoked it in the Control.Panels directory,
BASIC.SYSTEM tried to invoke the STARTUP DA as if it were a BASIC
file.

Avoid using that special name, and just rename the DA.