Eyes: Use more precise oval drawing logic. Fixes #83
Previously, the Eyes DA drew the ovals for the eyeballs using floating
point math to calculate the coordinates for a 36-sided polygon and a
wide pen. This was clunky code and looked bad when the window was
large.

This change ports the QuickDraw "InitOval" and "BumpOval" routines to
6502 code, and draws the ovals using those. The logic is to initialize
an oval struct with InitOval, then for each scan line call BumpOval to
update the struct, which provides the left/right edges.

To draw a hollow oval, two oval structs are initialized (inner and
outer) and the left/right edges are drawn separately.

This is actually *slower* when the window is very large, but there's
probably more opportunity to optimize, e.g. potentially using
PaintRect for each scan line rather than MoveTo/LineTo, or moving
the logic entirely into MGTK at some point. The appearance of the
ovals is much better, though.