DeskTop: Following delete, close affected windows. Fixes #654
If a folder is deleted, make sure any windows that are no longer
valid are closed.

This is done by probing *all* windows and closing any where the source
file is no longer reachable. This is simpler than trying to track the
information needed during the delete, and could prove handy in the
future. It has the side effect that all windows are probed, meaning
entirely unrelated windows will be tested and potentially closed,
including probing windows for open floppy disks (slow!).