DeskTop: Disallow creating a shortcut for a volume path copied to ramcard
DeskTop allows creating shortcuts to volumes and folders, but the
various places that do "Copy to RAMCard" (DeskTop.system, Selector,
DeskTop) have logic that doesn't handle volume paths, since they try
to copy a parent folder.

Since this is a very obscure case, rather than inventing behavior,
just block it. The alert isn't very helpful in this case, but it's not
worth adding a more specific string for this edge case.

An alternative would be to block shortcuts for volume paths, but
they're helpful for quick access within DeskTop itself.

Fixes #651. Fixes #652. Fixes #653.