File Picker: Clear to right of IP when changing drive.
The file picker inputs work by maintaining a left-of-IP string and a
IP-and-right-of-IP string. When Change Drive was clicked, the left
string was updated to the new volume's path, but the right was not
updated, leading to e.g. "/Vol|ic.System".

Clear the right side of the string, and factor out a helper for it.