Control Panel DA: Pull "active pattern" out of cycle.
For the desktop pattern selection, the "active pattern" took the place
of the default checkerboard in the selection sequence, until you
looped or backtracked e.g. click right then left.

Change this so that the active pattern exists outside of the sequence,
so if you click right you'll see the checkerboard. After clicking the
left or right arrow the "active pattern" no longer appears. This
matches the classic macOS control panel behavior.

This is related to issue #453 in that the "active pattern" shouldn't
replace one of the patterns until you moved away and back, but it
doesn't implement the suggestion that the "active pattern" should be
represented in the cycle.