Control Panels: Prompt/retry if system disk not present
The Control Panel DAs (Control Panel, Startup Options, Date and Time)
had logic to write state back to DeskTop at the the current prefix and
then, if this was a RAMCard, try to write to the original location and
offer to retry ("Please insert the system disk") if that failed. What
they didn't do was offer to retry if the initial write failed.

Move the prompt/retry logic so that it is used for both the initial
write, and (if that succeeded and is necessary) to the second write to
the original location.