DeskTop: Restrict Format/Erase dialog to 12 volumes. Resolves #326
Hacky fix for a niche case. The Format/Erase dialog only has room for
12 devices (3 columns, 4 rows). The absolute maximum number of devices
supported by DeskTop is 13 (7 slots * 2 drives, minus S3D2 for /RAM),
which is too many. The result would be that the top right corner would
contain two overlapping volume names. Wackiness ensues.

Extending the dialog doesn't seem worth it for this rare case.

Prevent this by excluding the last device in DEVLST from the list if
there are indeed 13 devices. Since this is now the startup volume, on
a system with so many devices this is rarely the one that would be
formatted anyway.