DeskTop: Fix regression with clearing selection.
A local error state was previously initialized to 0 on start, and
following the space optimizations in 4c330161 was initialized to a
non-zero dummy value. The code erroneously relied on the value still
being 0, or set to 0 when a window's selection was cleared, but if
this wasn't true then volume icons wouldn't be unhighlighted.

Force initializing it to 0, like it always should have been.