DeskTop: Move startup vol to start of device list. Resolves #468
DeskTop previously (well, as of 1.2) showed devices in reverse-DEVLST
order, likely putting the boot device at the top right. This is mostly
fine, but... I'm using a RAM disk driver that puts itself last in the
device list. The preferred order for ProDOS devices is to put Disk II
drives first, other physical drives next, and RAM drives last because
resolving pathnames to devices is done from last to first. So you want
fast RAM disks before slow spinning disks. And I like having my HD in
the top right.

So... instead, move the volume DeskTop is running from to the end of
the list during DeskTop startup so that it shows in the top right of
the screen. If DeskTop is copied to a RAM disk, the RAM disk will show
there. The order is restored when DeskTop exits.

This doesn't affect the device order seen in Disk Copy or Selector.