DeskTop: Make startup volume first, even if running from RAMCard.
This is a refinement of f3dba83e for #468, as suggested by #658.

As @c0mmander8 points out, even when DeskTop has copied itself to
RAMCard for faster execution, the user likely doesn't want to operate
on files from that volume by default e.g. in dialogs. So in that case,
check to see if the original startup volume is still online, and if so
move that volume to the start of the list.