Fit & Finish: Update modal dialog frame style
After further consideration from 79b64ddf, update the style used for
modal dialogs so that the have a non-XOR inner border. Update this to
be 4x2 DHR pixels rather than 2x2, which gives it a more consistent
thickness (since DHR pixels are close to 2:1), and change the spacing
from the dialog frame to also be 4x2 which seems to be visually
pleasing.

All modal dialogs are updated to match. Plus this fit and finish:

* Selector's buttons are drawn with 1x1 pen (was 2x1). Alert borders
* now have better spacing. Radio button/checkbox hit targets
* dynamically computed for: DeskTop's Edit Shortcut dialog Startup
* Options control panel DeskTop's shortcut picker dialog left column
* position and highlight improved. Handful of other slight UI layout
* tweaks.

Introduce docs/HIG.md to document interface guidelines.