DeskTop: Implement type-down selection. For #275
If unmodified filename characters are typed, select the matching icon.
The match is case-insensitive and matches modern macOS behavior: The
first icon in the active window or desktop icon that sorts
lexicographically at or after the typed string is selected, or the
last icon if there isn't one.

Any other event (e.g. mouse action or modified character) will cause
the typedown sequence to be reset.

Candidate icons are file icons in in the active window plus volume
icons on the desktop, just like the icons selectable with arrows.