DeskTop: Fix Special > Get Size showing 0k for normal files
The root cause was that SizeOrCountProcessSelectedFile would call into
SizeOrCountProcessDirectoryEntry to call GET_FILE_INFO for individual
selected files. The latter would try to compose a path using the
"current" FileEntry. This was bogus since the FileEntry was never
initialized in this path, since it was iterating the selection, not a
directory. But since the FileEntry buffer was initialized to 0, it
would be interpreted as a zero-length string and "just work". At
least, at first, until an actual directory was processed. So it was
always subtly broken.

But this was made more apparent by the following changes:

* 97249c32 moved the current FileEntry from a dedicated, 0-initialized
  buffer to page $0C which happened to be unused during operations.
  But $0C is used in plenty of other places, so it would normally
  contain garbage data, leading to a bogus path, so GET_FILE_INFO
  would fail, leading to reporting 0k.

* 5511e1af subsequently dropped a check for 0-length paths when
  concatenating. So that meant that even if the change in 97249c32 was
  reverted, the path composed for GET_FILE_INFO would be bogus.

The fix is:

* Populate the FileEntry buffer's name length to 0 when delegating
  from SizeOrCountProcessSelectedFile into
  SizeOrCountProcessDirectoryEntry.

* Add a check for the name length before concatenating. The check is
  not restored to the AppendFileEntryToSrcPath proc, though, because
  this is the only place this should happen.