DeskTop: Optimize table lookups and fix edge case in previous
Rather than ldx window_id / dex / lda table,x, skip the decrement and
use table-1,x

This revealed a bug in the last commit's use of FindIconByName; while
it works for volume icons, it was trying to check the
win_view_by_table entry for the desktop (window=0) so... looking at
bogus memory. Which happened to work when doing the decrement. Skip
the view-by check for the desktop, which is always icons.