DeskTop: Improve mouse responsiveness.
InitPort/SetPort was being called every tick of the main loop. Prior
to the recent SetZP changes this made the mouse much more janky in
DeskTop than when a dialog was showing, at least at 1MHz. But even
after the SetZP changes, there was still a noticeable difference.

Just drop this call. Anything that was relying on the ports being set
up this way was already fragile, so will be fixed as problems are
noticed.