Build: Factor out the "what to install" vs. "how to install it"
There are four installation systems:
  * bin/package -- produces distributable disk images (.2mg, .hdv, .po)
  * bin/shk ------ produces a ShrinkIt file
  * bin/install -- installs to an existing disk image
  * bin/mount ---- installs to a directory for mounting with Virtual ][

All four of these had duplicated logic for e.g. enumerating the desk
accessories, creating folders, and so on. This meant that when adding
a new file (e.g. OPTIONAL/UNSHRINK), all four files needed to be
touched.

Factor out the logic for what to install into bin/manifest, which
outputs a list of instructions for what to install. The installers
then execute these as a series of steps (add_file, create_folder,
remove_file), with logic specific to what they're operating on.