Don't hide Selector file; just don't run it by default.
The old behavior is that if SELECTOR was present in the same directory
as DESKTOP.SYSTEM then SELECTOR would be launched instead. This
allowed two scenarios:

* Users could decide to put SELECTOR in place so it loads as a quick
  app launcher, without starting all of DeskTop. Users could still run
  desktop via the button.

* Admins (usually teachers) could configure SELECTOR for users, and
  remove DeskTop entirely. This would hide the DeskTop button in
  Selector, making it harder to mess up the system.

This made sense "back in the day", before config files. Now that the
Startup Options control panel allows easily toggling whether Selector
starts, requiring filesystem changes to enable Selector in the first
place makes an unused feature even more difficult for users.

This commit changes things so SELECTOR is always placed in the main
directory, and the default option is changed to skip it. Users can go
into Startup Options and turn Selector on to have it start instead of
DeskTop.

Admins can still remove DESKTOP2 to get the previous behavior.

And to round out the change:

* The OPTIONAL/ directory is renamed to EXTRAS/

* For developers (like me), the "make installsel" option no longer
  makes sense. Instead, a "bin/setopt" utility is added which will
  modify an existing install's config file via Cadius to turn on or
  off the "Run Selector" option from the command line, passing "sel"
  or "nosel" as arguments.