DeskTop: Make Push/PopPointers routines preserve A,X
Desktop reuses $06/$07 and $08/09 on the zero page as 16-bit pointers
in many, many routines. Since there are inevitably conflicts between
routines, there's a PushPointers proc that spills these onto the
stack, and a corresponding PopPointers. This way, a caller can
preserve its pointers when it is calling a routine that will trash
them.

The routines would trash A,X, which is unfortunate. That meant they
couldn't be used easily in routines that took A,X as arguments without
saving the arguments first. This change makes the routines preserve
through the call (not pushed to the stack!), allowing the callee to do
the preservation easily, e.g.:

.proc Foo
    jsr     PushPointers
    stax    $06 ; A,X were preserved
    ...
    jsr     PopPointers
    rts
.endproc

This is close to bytes-neutral since many callers preserved A in other
ways.