DeskTop: Dedupe PushPointers/PopPointers
Main and Auxmem had their own copies of the Push/PopPointers procs.
Dedupe these by moving them into the handy LC shared proc space. in
$D000-$D200.

This couldn't be done prior to a0bf085c because GetIconInfo was called
with LCBANK2 paged in unnecessarily. That was fixed as part of the
cleanup there, so the deduping is unblocked.