Sort Directory: Use DeskTop's 10k buffer to eliminate limits. Fixes #86
Sort Directory was limited to 7 blocks or ~90 entries, because it used
the DA memory area only.

Switch to the 10k dynamic buffer in DeskTop (same as the Duet player),
which allows up to ~260 entries. Which is way more than needed,
because DeskTop itself only handles 127 icons total, so after you take
away all your drives and trash drives you're likely limited to 113 -
124 entries in a visible directory anyway.