This Apple DA: Detect Uthernet II card in Slot 3
Add a special case for slot 3. It should always detect an 80-column
card, because desktop doesn't support running without one... but it's
actually in the Aux slot and some cards w/o firmware like the Uthernet
II can peacefully reside in the real slot 3.

After checking slot 3 the normal way, also test for an Uthernet II
and tack on ", Uthernet II" if found.