New DA: Sounds control panel. Resolves #283
This is a new Control Panel DA that allows selecting from a set of
audio routines. The routines must fit in 128 bytes and are played back
from $300. They can use $00-$05 as ZP locations, and interrupts are
disabled during playback.

The different apps (DeskTop, Selector, Disk Copy) load the file at the
same time as settings, and stash it for playback. The Bell proc
previously used to generate the ProDOS "buzz" tone now relocates the
sound proc to $300 (a memory swap), plays the tone, and restores $300.

The sounds include the ProDOS "buzz", the classic Control-G "beep",
the IIgs "bonk" (Michael J. Mahon's version), and a variety from other
Apple II applications and Apple Assembly Lines. It's easy to add more.

(The Karateka "hya!" sound is too large to include, unfortunately.)