Add ROMCall lib, use it in Calculator and Sci.Calc to simplify.
The routines in the Calculator DAs made numerous calls into ROM
with wrapper procs for each to save processor flags and bank
switch.

Replace this with an MLI-call-style helper that looks up the
actual ROM address to call from the call site, and takes care
of register preservation and bank switching. In Sci.Calc
this saves 230 bytes.

Resolves #77