Fit & Finish: Use XOR-drawn insertion point in text fields
There are 4 text field implementations:
* DeskTop's name prompt (used in Rename, etc)
* Find File DA's search field
* File picker's two fields (target, target+shortcut name, or src+dst)

All four are implemented using a left/right buffer pair, split at the
insertion point (a.k.a. caret). The insertion point was implemented as
glyph, present at the start of the right buffer. This had several
drawbacks:

* Manipulating the strings required many off-by-one considerations.
* Moving the IP requires redrawing at least part of the string.
* Arrowing through the string would shift characters.

This change simplifies things by rendering the IP as an XOR-drawn line
on top of the text, rather than as a character in the string. This
greatly simplifies the string manipulations and improves the
appearance, and reduces the number of string redraws needed.

It does add back some complexity in that IP moves and string
manipulations need to carefully balance hiding/showing the IP so that
debris is not left behind.