File Pickers: Remove over-configurable jump table
In two-field mode, the file pickers supported letting the caller
configure (via a jump table) how to handle various actions, such
as various keyboard actions, clicks, and so on. This went unused
as the dialogs simply called the default actions for all of these.
This made adding new actions to the pickers onerous.

Remove this degree of configurability.