File Pickers: Resolve issues with changing paths and selection
If you selected a file so that the path was e.g. "/Volume/File", put
the put the insertion point in the middle of the path like
"/Vol|ume/File", and clicked "Another", the path would be updated to
be "/VolAnother" or something equally bogus.

The path input fields would also be redrawn three (!!!) times whenever
list selection was changed. There's also some weird "dirty state"
tracking that alters some picker behavior in subtle ways that I don't
fully understand, but it was also broken. And then there were about 8
different "get the Nth filename" implementations.

This commit cleans as much of that up as I could find and understand.
It's likely that some edge cases were missed, but I have great faith
in our highly paid QA team.