Alerts: Use a different bitmap when there are no choices.
MouseDesk/A2D uses a very early Susan Kare bitmap for alerts, a face
with a question mark. In early iterations of this visual style, there
were three similar icons: a face with a question mark, a face with an
exclamation mark, and a face with text. These didn't last long, being
replaced by an alert triangle or a stop hand. I really like this
quirky original style, so definitely keeping the face.

That said, the question mark only makes sense if there are options. So
introduce the exclamation mark bitmap, and use that if the dialog only
has an OK button. If there are multiple options, the question mark
version is shown.

The alert procs are also tweaked to play the bell after the dialog is
drawn, which prevents the sound from slowing the drawing for
accelerators (or emulators) that slow down when the speaker is
touched.