DeskTop: Make sure filename is retained in Copy a File dialog.
The recent File Picker / insertion point fixups regressed some
functionality. When second path was updated via a selection (or
open/close/change drive), the filename suffix was lost.

Previously this was done via a very fragile approach - the text after
the insertion point was retained. This was fine unless the IP was
moved. The fragility was addressed, but the functionality was lost.

Restore this by always copying the last segment; this "just works" if
you don't move the IP or type anything, and is resilient to the IP
moving. If you actually type a new destination name then click around,
the new name will be preserved and appended to the newly selected
path.