DeskTop: Update icon following a file rename. Fixes #574
Some files are given icon based on the name, e.g. SYS files are
assigned different icons depending on whether or not the name has the
".SYSTEM" suffix. The same now applies to ".SHK" and other archive
types. This makes the problem with renames requiring a window reload
more acute; SYS files were special cased, but archive types were not
(and could be any file type - BIN, TXT, etc).

So have the rename logic potentially replace the icon details when a
file icon is renamed. This handles the SYS rename problem and also
handles archives.