DeskTop: Don't animate window opening with bogus source
The logic to animate window opening assumed that window was associated
with an icon, which used to be true. But since windows can now be
opened from paths (e.g. Apple menu, shortcuts, or restore) this wasn't
always true. So the code would animate... from something.

Instead, just skip the animation if there's no source icon.