File Dialog: Don't blindly copy strings to LC buffer
In DeskTop, where app logic is in main but MGTK is in aux,
runtime-modified strings need to be in the LC bank. The file dialog's
DrawString would copy every string being drawn to an LC buffer to
support this.

But most of the strings are static resources, so this is not
necessary. Most of the dynamic string are also in LC buffers. The
exception is volume and filenames - so just copy those into an LC
buffer, and simplify string drawing.

Also, add a missing RTS at the end of a proc. Got lucky with that one.