Line Edit: Replace "right buffer" with just an IP position
The Line Edit logic was implemented using a left/right buffer
algorithm. The IP is implicitly between them. Inserting a character
simply appends to the left buffer, which is fast. Moving the IP
requires shuffling characters around, though, and there is a lot of
other complexity e.g. when clicking, separate code paths depending on
clicks to the left or right of the IP.

Simplify this down to having a single string buffer, and maintain an
IP position. Since strings are short (<=64 characters) this is never a
performance problem and the code savings are substantial -- over a
kilobyte, but spread out in overlays so not that useful.