File Dialog: Simplify GrafPort usage and remove unnecessary switches
The original A2D code seems to reset GrafPorts aggressively back to
the "default" when work is complete. But a better style is to never
assume the GrafPort you want is selected and explicitly set it before
using it. The File Dialog code does do proper port selection, so
remove the resets.

Also, since there are only two GrafPorts used, optimize the calls.

This allows some size reduction in file segment sizes, now more
intelligently applied to Selector. Between this and previous
optimizations, a whole disk block is freed up!