Desktop: Fix repainting of overlapping icons after dragging. Fixes #687
The IconTK's DragHighlightedImpl's loop for moving icons within a
window calls erase_icon to erase each icon. But that redraws
overlapping icons. So if the dragged icons overlap each other, they
are accidentally redrawn.

Fix by introducing a flag for erase_icon which causes
RedrawIconsAfterErase to not redraw highlighted icons.