DeskTop + Selector: Eliminate most unnecessary GrafPort switches
Like previous commit for File Dialog: The original A2D code seems to
reset GrafPorts aggressively back to the "default" when work is
complete. But a better style is to never assume the GrafPort you want
is selected and explicitly set it before using it.

Before drawing to the desktop (open/close animation, volume icons)
it's important to init/set a desktop port. The only gotcha is that
after a DA window closes the DA's port is invalid, whereas desktop's
windows (directory windows, dialogs, even dialogs in overlays) are all
in LC space and technically still valid even if the window is gone.
Code like the clock will save/restore the port, and restoring an
invalid port corrupts memory, so the port needs to be reset after a DA
exits, too.

Saves 89 bytes in desktop/main and probably improves performance