Shave 23 bytes in desktop/main by simplifying handling of window clicks
Code inspection reveals that there are several code paths where
control returns the main loop with a window's icons cached. For
example, clicking on a window's header. Most code is written to always
swap the desktop's icons back into the cache after futzing with a
window's icons. If that's necessary, there are bugs. If that's not
necessary, there are a lot of extraneous calls. For now, just note
cases with comments.