DeskTop: Fix copying flags/prevent corruption when deleting shortcuts.
The logic for copying the "copy to ramcard?" flags when moving
shortcuts in the "secondary run list" (those that don't show in the
menu) was flawed, done with the wrong pointers. This led to file
corruption.

Found by code inspection while looking for duplicate code, but
easy to repro. Fixed, also deduped the code.