MGTK: Prevent screen artifacts on startup
Eagle-eyed users would notice that on certain systems during startup,
while the desktop background was being drawn there would be two stray
pixels at the bottom of the screen for a moment. Mysteriously, this
seemed to happen during a call into the mouse firmware! Thanks to
@peterferrie for figuring out the trigger.

The Apple II Mouse Card firmware inspects the F8 ROM signature byte
$FBB3 for startup, and on systems earlier than the IIe it does modify
the hi-res screen, in order to establish vertical sync. This is
documented in the manual! Of course, when MGTK is initializing in
DeskTop, the ROM is not banked in because we run MGTK from Aux and we
keep a lot of resources in LC. So when the mouse firmware looked at
$FBB3 it saw a value other than $06 and ran the II/II+ initialization
code, which displayed and dirtied the screen.

Other code in MGTK already wrapped a SETMOUSE call with setting the
RAM byte $FBB3 temporarily to $06; we just need to apply the same
logic to INITMOUSE and the problem goes away. Yay!

Aside: Why is the hi-res screen even shown before the desktop
background is done painting? MGTK uses the text pages as a pattern
buffer, so the text pages are dirty when the background is being
shown. Something needs to be visible! If we hadn't found this fix,
other alternatives would have been to reorder MGTK's StartDeskTop
sequence to init the mouse earlier.