Show Text File DA: Scrollbar improvements. Fixes #422
* Change keys to match other scrollable areas:
  * Up/Down - Line up/down
  * Apple+Up/Down - Page up/down
  * OA+SA+Up/Down - Home/End
* Conditionally activate scrollbar
* Move scrollbar proportionally through the document
* Reset scroll position when toggling Proportional/Fixed display

Moving up/down by a single line per scroll tick has the consequence
that you won't be able to scroll through files greater than 255 lines.
This can potentially be improved in the future by not requiring an
integral scroll tick per line, but we'll see if this is a problem in
practice.

Positioning the scrollbar proportionally requires a heuristic. After
the first page of text is displayed, a calculation is made about the
fraction of the document shown. If the whole file was shown, the
scrollbar is deactiated. Otherwise, an estimate of the total number of
lines in the file is made to set the maximum scroll value. In
practice, this seems to work, but there are going to be some pedantic
cases where it doesn't. Still, this feels like an improvement, so we
can iterate from here.