Line Edit: Impose a clipping rect to prevent overflowing.
A full 64-character pathname would overflow the file dialog. While
uncommon, this is ugly. But making the file dialog even larger would
be worse since it's already a beast. And there would need to be room
for the spaces used to erase the end of a string.

So... reintroduce the concept of a cliprect, and clip all drawing into
the line edit control. This does bump up the byte count of the file
dialog impls back up a bit, but we've earned it.