File Dialog: If building in "simple" config, make dialog shorter.
For the build configuration used in the Selector app, there's no need
to leave room for a second text edit control. This is only used in
DeskTop, for Copy a File and Add/Edit a Shortcut. So the dialog can be
shorter.

This is done at build time, though, so the space is still present
but empty for Delete a File.