File Dialog: Display disable navigation buttons as dimmed
The Change Drive, Open and Close buttons navigate the disk/list. The
behavior of these is somewhat subtle and tied to navigating the file
list. In some cases (e.g. Add/Edit a Shortcut) they are entirely
disabled when the shortcut title is being edited. The Change Drive
button is a no-op when there is only one device. The Close button is
disabled until you've opened a folder, and the Open button is disabled
unless a folder is selected. But there was no visual feedback for
this.

Introduce the ability to "dim" the buttons (gray out the label) and
dim them when disabled.