DeskTop: Ensure correct handling after an error in File > Delete a File...
A missing RTS in the error handler made bad things happen.

Also ensure that an OK in Selector's File > Run a Program... with no
selection gives some feedback rather that silently ignoring the
button. Which was what I set out to fix. The bug above was discovered
via code inspection when aligning the implementations.