File Dialog: Set ports and avoid flickering dimmed buttons
Button re-drawing on click was not setting a GrafPort. This was not
typically a problem as the line edit would have set a port, but this
shouldn't be relied upon.

Also, when updating the dimmed button state, even if the state didn't
change the button labels would be redrawn resulting in a flicker. Add
flags to track the dimmed state for each button to avoid unnecessary
redraws.