DeskTop: Tweak RAMDisk icon.
Give it a hint of a slant to fit under the keyboard in a IIe, like a
RAMWorks card. It's not a perfect likeness; in reality the onnector
would be centered. But maybe you're using Apple's Slinky card,
although that doesn't have a slant. I dunno. Anyway, it adds a little
character and was voted higher than the existing one.

It continues to look a bit like a derpy skull to me...

.            #############################
.         ###                           ##
.      ###                      #####   ##
.   ###                #####   ##   ##  ##
. ##                  ##   ##  ##   ##  ##
. ##                  ##   ##  ##   ##  ##
. ##                  ##   ##  ##   ##  ##
. ##                   #####    #####   ##
. ##                                    ##
. ########################  ##  ##  ##  ##
.                       ##  ##  ##  ##  ##
.                       ##################