DeskTop: Drop assumption that the desktop icon tables are cached.
Each window plus the desktop has a half-page in Aux that lists count
and number of entries it contains, in presentation order. For icon
view, each entry is an icon number, in disk order. For other views,
each entry is an index into the window's FileEntry table, in sorted
order. These are updated when files are added/removed, or when the
view is rebuilt. Since these reside in Aux, they're inconvenient to
access. So there are helper routines in LC to cache a table in LC
memory, and to persist the cache back.

Most of DeskTop was coded assuming that the desktop's table is cached.
Routines that operated on e.g. the active window would load it into
the cache, optionally store it back if changes were made, and reload
the desktop into the cache. This was (1) extra work (code
size/performance) and (2) fragile, since every code path in a proc
that cached a different window needed to exit carefully to ensure the
desktop was re-cached. But not every path did this correctly, despite
contortions (e.g. handling clicks on windows and drag/drop).

Since this assumption was already violated, rather than trying to fix
all of the cases, just drop it. Require anything that operates on the
desktop's icons to cache them in, and don't assume any particular
state. This makes verifying correct behavior more local. Any bugs that
are revealed by this change were probably pre-existing, it's just
easier to reproduce them now.

Saves 87 bytes in main.